ensure gform key is set before disabling installation screens

+++ genero-adjustments.php
@@ -91,8 +91,10 @@ class Adjustments {
       add_filter('gform_enable_field_label_visibility_settings', '__return_true');
 
       // Gravityform forces these.
-      delete_option('gform_pending_installation');
-      delete_option('gform_enable_background_updates');
+      if (get_option('rg_gforms_key') && get_option('gform_pending_installation')) {
+        delete_option('gform_pending_installation');
+        delete_option('gform_enable_background_updates');
+      }
     }
   }
